Tryed to remove black border in EquirectangularAdapter if blur enabled
This commit changes the image rendering logic in the EquirectangularAdapter class.
If the config.blur flag is set,  adds a margin to avoid black borders.
Otherwise, it draws the image as previous.
This change is not the best solution possible but for me it's more annoying if I see a black line than cropping a fiew pixels.

If you try, it's not really easy to notice this small crop.

I basically apply a margin of 0.1% calculated over the width of the panorama so as to hide the shaded border.